requirements.txt
asgiref==3.3.1
Django==1.11.17
djangorestframework==3.7.7
Pygments==2.7.2
PyMySQL==0.10.1
python-memcached==1.59
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1